refactor(tests): example file handling in a module
